chore (api|client): Add node-version to src directories
In some circumstances it might be necessary to install the npm packages
locally. For that case we provide a `.node-version` file. When using
node version managers developers can easily bring up the required
version. That way they are working with exact the same version as
expected in the container.

**Example for fast node manager**

```bash
cd ./app/client
fnm install
fnm use
fnm current
```

``